feat(sway)!: automatic colorscheme updating on sway reload
The colorscheme filename no longer changes in the sway config.  Instead
this file is a symbolic link to a colorscheme.  Consequently, reloading
the sway config will read the file again and apply the colorscheme.

Updating of this file happens when setting a new base16-shell theme in
the zshrc file and then running exec $SHELL.  The exec action updates
the symbolic link to reference the new chosen theme.

A possible future improvement is to wrap the action in a shell function
eliminating the need to exec or restart the shell.